refactor server module into server.rs
